backend/src/main/java/com/popflix/model/Watchlist.java
package com.popflix.model;

import org.springframework.data.annotation.Id;
import lombok.Data;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.NoArgsConstructor;

@Builder
@NoArgsConstructor
@AllArgsConstructor
@Data
public class Watchlist {
    @Id
    private Long id;
    private Long userid;
    private Integer movieid;

    public Watchlist(Long userid, Integer movieid) {
        this.userid = userid;
        this.movieid = movieid;
    }

    public Long getId() {
        return this.id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getUserid() {
        return this.userid;
    }

    public void setUserid(Long userid) {
        this.userid = userid;
    }

    public Integer getMovieid() {
        return this.movieid;
    }

    public void setMovieid(Integer movieid) {
        this.movieid = movieid;
    }
}